requirements.txt
alembic==1.8.0
Babel==2.9.0
black==22.3.0
click==8.1.3
colorama==0.4.4
config==0.5.1
distlib==0.3.4
filelock==3.7.1
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.37
tomli==2.0.1
Werkzeug==2.1.2
WTForms==3.0.1
